+++ packages/webpack-config/src/bundles/getBundleEntry.ts
@@ -1,9 +1,8 @@
 import { getEntry } from "../common/getEntry";
 
-const getBundleEntry = (name: string, bundle: boolean): unknown => {
-    const fixName = name && name.startsWith(".") ? name.substring(1) : name ? name : "";
-
-    return getEntry("", fixName, bundle);
-};
+const skipFirstDotPosition = 1,
+    getFixedName = (name: string): string =>
+        name && name.startsWith(".") ? name.substring(skipFirstDotPosition) : name || "",
+    getBundleEntry = (name: string, bundle: boolean): unknown => getEntry("", getFixedName(name), bundle);
 
 export { getBundleEntry };