+++ packages/eslint-config/CHANGELOG.md
@@ -3,6 +3,51 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.4](https://github.com/tsparticles/utils/compare/v2.0.3...v2.0.4) (2023-12-13)
+
+**Note:** Version bump only for package @tsparticles/eslint-config
+
+
+
+
+
+## [2.0.3](https://github.com/tsparticles/utils/compare/v2.0.2...v2.0.3) (2023-12-13)
+
+**Note:** Version bump only for package @tsparticles/eslint-config
+
+
+
+
+
+## [2.0.2](https://github.com/tsparticles/utils/compare/v2.0.1...v2.0.2) (2023-12-13)
+
+**Note:** Version bump only for package @tsparticles/eslint-config
+
+
+
+
+
+## 2.0.1 (2023-12-12)
+
+
+### Bug Fixes
+
+* **deps:** update dependency eslint-config-prettier to v9 ([b950d9f](https://github.com/tsparticles/utils/commit/b950d9f2eeca55228d613d6db00ffba0402d0634))
+* **deps:** update dependency eslint-plugin-jsdoc to v44 ([b00f1fd](https://github.com/tsparticles/utils/commit/b00f1fd85756fb21c0eb3625f032d27fc1332659))
+* **deps:** update dependency eslint-plugin-jsdoc to v45 ([4d0e75c](https://github.com/tsparticles/utils/commit/4d0e75c42cbde77cf5413c1322423b6e04085d09))
+* **deps:** update dependency eslint-plugin-jsdoc to v46 ([91247e7](https://github.com/tsparticles/utils/commit/91247e794b3d8275afe12740fdd178a8e69b8b75))
+
+
+### Features
+
+* added additional externals parameter to webpack plugin ([01c94e8](https://github.com/tsparticles/utils/commit/01c94e8aea203c6c277cc612848a2b22a928a230))
+* added effects support ([d1e1743](https://github.com/tsparticles/utils/commit/d1e17431a2b1af081f62f0e52bc7436e3b83e863))
+* added eslint-plugin-import for enforcing .js at the end of every import ([be58415](https://github.com/tsparticles/utils/commit/be5841584f4b3a6d6b011031548e8c1929b05c2c))
+
+
+
+
+
 # [2.0.0](https://github.com/tsparticles/utils/compare/@tsparticles/eslint-config@2.0.0-beta.8...@tsparticles/eslint-config@2.0.0) (2023-11-21)
 
 **Note:** Version bump only for package @tsparticles/eslint-config