+++ packages/webpack-config/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.4](https://github.com/tsparticles/utils/compare/v2.0.3...v2.0.4) (2023-12-13)
+
+**Note:** Version bump only for package @tsparticles/webpack-plugin
+
+
+
+
+
+## [2.0.3](https://github.com/tsparticles/utils/compare/v2.0.2...v2.0.3) (2023-12-13)
+
+**Note:** Version bump only for package @tsparticles/webpack-plugin
+
+
+
+
+
+## [2.0.2](https://github.com/tsparticles/utils/compare/v2.0.1...v2.0.2) (2023-12-13)
+
+**Note:** Version bump only for package @tsparticles/webpack-plugin
+
+
+
+
+
+## 2.0.1 (2023-12-12)
+
+
+### Bug Fixes
+
+* **deps:** update dependency eslint-config-prettier to v9 ([b950d9f](https://github.com/tsparticles/utils/commit/b950d9f2eeca55228d613d6db00ffba0402d0634))
+
+
+### Features
+
+* added additional externals parameter to webpack plugin ([01c94e8](https://github.com/tsparticles/utils/commit/01c94e8aea203c6c277cc612848a2b22a928a230))
+* added effects support ([d1e1743](https://github.com/tsparticles/utils/commit/d1e17431a2b1af081f62f0e52bc7436e3b83e863))
+* added version to webpack output ([496fcbb](https://github.com/tsparticles/utils/commit/496fcbb98c64cdb37c874bfbf8f9c3172806ef58))
+
+
+
+
+
 # [2.0.0](https://github.com/tsparticles/utils/compare/@tsparticles/webpack-plugin@2.0.0-beta.14...@tsparticles/webpack-plugin@2.0.0) (2023-11-21)
 
 **Note:** Version bump only for package @tsparticles/webpack-plugin